refactor: split hooks.ts into hooks.ts and hooks-types.ts
- hooks-types.ts: helper functions (executeHook, executeBeforeOperation)
- hooks.ts: exported run*Hooks functions
- README: updated internal files list
