+++ packages/collections/src/collections/hooks-types.ts
@@ -1,0 +1,35 @@
+// Hook executor types and helpers
+
+import type {
+  CollectionHooks,
+  HookHandler,
+  BaseHookContext,
+} from './types'
+
+/**
+ * Execute a hook handler, allowing both sync and async return
+ */
+const executeHook = async <T extends BaseHookContext>(
+  handler: HookHandler<T> | undefined,
+  context: T
+): Promise<T> => {
+  if (!handler) return context
+
+  const result = handler(context)
+
+  // Support both Promise and sync return
+  return result instanceof Promise ? result : result
+}
+
+/**
+ * Execute before operation hooks
+ */
+const executeBeforeOperation = async (
+  hooks: CollectionHooks,
+  context: BaseHookContext
+): Promise<BaseHookContext> => {
+  const ctx = await executeHook(hooks.beforeOperation, context)
+  return ctx
+}
+
+export { executeHook, executeBeforeOperation }
